content/base/test/test_bug340571.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=340571
-->
<head>
  <title>Test for Bug 340571</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=340571">Mozilla Bug 340571</a>
<p id="display"></p>
<div><select id="sel"></select></div>
<script type="application/javascript">
ok(!document.getBoxObjectFor, "Detecting getBoxObjectFor");
try {
  document.getBoxObjectFor(document.body);
  ok(false, "getBoxObjectFor succeeded");
} catch (e) {
  ok(true, "getBoxObjectFor succeeded");
}

var sel = document.getElementById("sel");
ok(!sel.boxObject, "Getting boxObject");
</script>
</body>
</html>
